1. landing page 2. basic form validation 3. using localStorage for filename
